backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==1.10.13
python-multipart==0.0.6
PyPDF2==3.0.1
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
openai==1.3.5
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
aiosqlite==0.19.0
python-docx==1.1.0
whisper==1.1.10
opencv-python==4.8.1.78
moviepy==1.0.3
torch==2.1.1
transformers==4.35.2